+++ OMG.PSUtilities.Core/OMG.PSUtilities.Core.psm1
@@ -32,6 +32,7 @@ $PublicFunctions = @(
     'Get-PSUModule',
     'Get-PSUUserEnvironmentVariable',
     'Get-PSUUserSession',
+    'Get-PublicIP',
     'New-PSUGithubPullRequest',
     'New-PSUHTMLReport',
     'New-PSUOutlookMeeting',